Day 04 completed
Converts a bingo card to a prop with keys as entries and values as their
locations, so that when calling numbers we can use the prop to check
which row and column that number lies on.  Then incrementing a counter
for each row and column makes it simple to check when a line is made.